requirements.txt
tensorflow>=2.10.0
streamlit>=1.20.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
gradio>=3.0